requirements.txt
flask
flask-cors
opencv-python
numpy
pillow